+++ mesos/zookeeper.go
@@ -3,6 +3,8 @@ package mesos
 import (
 	"encoding/json"
 	"strconv"
+	"strings"
+	"sync/atomic"
 
 	mesosproto "../proto"
 
@@ -42,6 +44,8 @@ func StatusZookeeper(id int) *cfg.State {
 
 // StartZookeeper is starting a zookeeper container with the given IDs
 func StartZookeeper(id int) {
+	newTaskID := atomic.AddUint64(&config.TaskID, 1)
+
 	var cmd cfg.Command
 
 	// before we will start a new zookeeper, we should be sure its not already running
@@ -61,25 +65,54 @@ func StartZookeeper(id int) {
 		}
 	}
 
+	networkIsolator := "weave"
+	var hostport, containerport uint32
+	hostport = 31210 + uint32(newTaskID)
+	containerport = 2181
+	protocol := "tcp"
+
+	cmd.TaskID = newTaskID
+
 	cmd.ContainerType = "DOCKER"
-	cmd.ContainerImage = "zookeeper"
+	cmd.ContainerImage = config.ImageZookeeper
 	cmd.NetworkMode = "bridge"
+	cmd.NetworkInfo = []*mesosproto.NetworkInfo{{
+		Name: &networkIsolator,
+	}}
+	cmd.DockerPortMappings = []*mesosproto.ContainerInfo_DockerInfo_PortMapping{{
+		HostPort:      &hostport,
+		ContainerPort: &containerport,
+		Protocol:      &protocol,
+	}}
+
 	cmd.Shell = false
-	cmd.TaskName = "Zookeeper" + strconv.Itoa(id)
+	cmd.TaskName = "av_zookeeper" + strconv.Itoa(id)
+	cmd.Hostname = "av_zookeeper" + strconv.Itoa(id) + config.ZookeeperCustomString + "." + config.Domain
 	cmd.InternalID = id
 	cmd.IsZookeeper = true
+	cmd.Volumes = []*mesosproto.Volume{
+		{
+			HostPath:      func() *string { x := config.VolumeZookeeper + "/" + strconv.Itoa(id); return &x }(),
+			ContainerPath: func() *string { x := "/datalog"; return &x }(),
+			Mode:          mesosproto.Volume_RW.Enum(),
+		},
+	}
 	sI := strconv.Itoa(id)
 	cmd.Environment.Variables = []*mesosproto.Environment_Variable{
 		{
+			Name:  func() *string { x := "SERVICE_NAME"; return &x }(),
+			Value: &cmd.TaskName,
+		},
+
+		{
 			Name:  func() *string { x := "ZOO_MY_ID"; return &x }(),
 			Value: &sI,
-		}, {
+		},
+		{
 			Name:  func() *string { x := "ZOO_SERVERS"; return &x }(),
 			Value: GetZookeeperServerString(id),
 		},
 	}
-
-	cmd.Hostname = "zookeeper" + strconv.Itoa(id) + "." + config.Domain
 
 	d, _ := json.Marshal(&cmd)
 	logrus.Debug("Scheduled Zookeeper: ", string(d))
@@ -92,7 +125,6 @@ func StartZookeeper(id int) {
 func initStartZookeeper() {
 	if config.ZookeeperCount <= (config.ZookeeperMax - 1) {
 		StartZookeeper(config.ZookeeperCount)
-
 		config.ZookeeperCount++
 	}
 }
@@ -106,7 +138,7 @@ func GetZookeeperServerString(id int) *string {
 		if i == id {
 			server += "server." + sI + "=0.0.0.0:2888:3888;2181 "
 		} else {
-			server += "server." + sI + "=zookeeper" + sI + "." + config.Domain + ":2888:3888;2181 "
+			server += "server." + sI + "=av_zookeeper" + sI + config.ZookeeperCustomString + "." + config.Domain + ":2888:3888;2181 "
 		}
 	}
 
@@ -119,8 +151,11 @@ func CreateZookeeperServerString() {
 	var server string
 	for i := 0; i < max; i++ {
 		sI := strconv.Itoa(i)
-		server += "zookeeper" + sI + "." + config.Domain + ":2181, "
+		server += "av_zookeeper" + sI + config.ZookeeperCustomString + "." + config.Domain + ":2181,"
 	}
+
+	// remote last char
+	server = strings.TrimSuffix(server, ",")
 
 	config.ZookeeperServers = server
 }