+++ changelog.md
@@ -15,3 +15,8 @@
 - Add RestCall to kill a task
 - Add Authentication
 - Add Support to configure (non)SSL Support to Mesos
+- Add Custom Domain to Zookeeper and Kafka to match Consul DNS
+- Add Service Name ENV Variable to Zookeeper and Kafka to match Consul DNS
+- Add Call_Suppress to tell mesos it does not send us offers until we ask
+- Add default values for some env
+- Add custom image name via env